requirements.txt
Flask-WTF
Jinja2
Werkzeug
blinker
decorator
flup
gunicorn
configparser
cov-core
coverage
cpplint
pyparsing
pytest
sqlalchemy
flask-sqlalchemy
sqlalchemy-migrate
flask-login
flask-openid
wtforms